+++ src/components/DashboardFilters/TimePeriodFilter/index.tsx
@@ -1,0 +1,3 @@
+import TimePeriodFilter from './TimePeriodFilter-container';
+
+export default TimePeriodFilter;
