+++ src/components/DashboardFilters/WatchListFilter/index.tsx
@@ -1,0 +1,3 @@
+import WatchListFilter from './WatchListFilter-container';
+
+export default WatchListFilter;
